Move all exceptions to exceptions.pywa
