+++ ocp_resources/node_network_configuration_policy.py
@@ -3,6 +3,7 @@ import re
 from kubernetes.dynamic.exceptions import ConflictError
 
 from ocp_resources.constants import TIMEOUT_4MINUTES
+from ocp_resources.exceptions import NNCPConfigurationFailed
 from ocp_resources.node import Node
 from ocp_resources.node_network_configuration_enactment import (
     NodeNetworkConfigurationEnactment,
@@ -13,10 +14,6 @@ from timeout_sampler import TimeoutExpiredError, TimeoutSampler, TimeoutWatch
 
 IPV4_STR = "ipv4"
 IPV6_STR = "ipv6"
-
-
-class NNCPConfigurationFailed(Exception):
-    pass
 
 
 class NodeNetworkConfigurationPolicy(Resource):
